Separating functions into their own files
